LPS-97123 Namespacing text extractor api
