+++ modules/apps/asset/asset-auto-tagger-api/bnd.bnd
@@ -5,7 +5,7 @@ Export-Package:\
 	com.liferay.asset.auto.tagger,\
 	com.liferay.asset.auto.tagger.configuration,\
 	com.liferay.asset.auto.tagger.exception,\
-	com.liferay.asset.auto.tagger.extractor,\
 	com.liferay.asset.auto.tagger.model,\
 	com.liferay.asset.auto.tagger.service,\
-	com.liferay.asset.auto.tagger.service.persistence
+	com.liferay.asset.auto.tagger.service.persistence,\
+	com.liferay.asset.auto.tagger.text.extractor
