+++ modules/apps/asset/asset-auto-tagger-api/src/main/resources/com/liferay/asset/auto/tagger/text/extractor/packageinfo
