+++ modules/apps/blogs/blogs-web/src/main/java/com/liferay/blogs/web/internal/asset/auto/tagger/text/extractor/BlogsEntryTextExtractor.java
@@ -12,9 +12,9 @@
  * details.
  */
 
-package com.liferay.blogs.web.internal.asset.auto.tagger.extractor;
+package com.liferay.blogs.web.internal.asset.auto.tagger.text.extractor;
 
-import com.liferay.asset.auto.tagger.extractor.TextExtractor;
+import com.liferay.asset.auto.tagger.text.extractor.TextExtractor;
 import com.liferay.blogs.model.BlogsEntry;
 
 import java.util.Locale;
