+++ modules/apps/journal/journal-web/src/main/java/com/liferay/journal/web/internal/asset/auto/tagger/text/extractor/JournalArticleTextExtractor.java
@@ -12,9 +12,9 @@
  * details.
  */
 
-package com.liferay.journal.web.internal.asset.auto.tagger.extractor;
+package com.liferay.journal.web.internal.asset.auto.tagger.text.extractor;
 
-import com.liferay.asset.auto.tagger.extractor.TextExtractor;
+import com.liferay.asset.auto.tagger.text.extractor.TextExtractor;
 import com.liferay.dynamic.data.mapping.model.DDMStructure;
 import com.liferay.dynamic.data.mapping.service.DDMStructureLocalService;
 import com.liferay.dynamic.data.mapping.storage.DDMFormValues;
